+++ api/helpers/checkedOutModel.js
@@ -5,7 +5,9 @@ const db = knex(knexConfig.development);
 
 module.exports = {
   getCheckedOut,
-  getCheckedOutById
+  getCheckedOutById,
+  getBookById,
+  getAllBooks
 };
 
 function getCheckedOut(userId) {
@@ -64,4 +66,49 @@ function getCheckedOutById(checkedOutId) {
     .first();
 
   return items;
+}
+
+function getAllBooks() {
+  const books = db("books")
+    // .join("users as lenders", "books.userId", "lenders.userId")
+    .select(
+      "books.bookId as bookId",
+      "lenders.firstName as lender",
+      "lenders.userId as lenderId",
+      "lenders.latitude as latitude",
+      "lenders.latitude as longitude",
+      "books.title",
+      "books.authors",
+      "books.image",
+      "books.isbn",
+      "books.avgRating",
+      "books.description",
+      "books.available",
+      "books.checkOutRequest"
+    );
+
+  return books;
+}
+
+function getBookById(bookId) {
+  const books = db("books")
+    .join("users as lenders", "books.userId", "lenders.userId")
+    .select(
+      "lenders.firstName as lender",
+      "lenders.userId as lenderId",
+      "lenders.latitude as latitude",
+      "lenders.latitude as longitude",
+      "books.bookId as bookId",
+      "books.title",
+      "books.authors",
+      "books.image",
+      "books.isbn",
+      "books.avgRating",
+      "books.description",
+      "books.available",
+      "books.checkOutRequest"
+    )
+    .where("books.bookId", bookId);
+
+  return books;
 }
