+++ api/routes/userRoutes.js
@@ -2,7 +2,8 @@ const express = require("express");
 const router = express.Router();
 const CheckedOut = require("../helpers/checkedOutModel");
 const Reviews = require("../helpers/reviewsModel");
-const Books = require("../helpers/bookModel");
+// const Books = require("../helpers/bookModel");
+
 const { authenticate } = require("../auth/authenticate");
 
 const db = require("../../data/dbConfig");
@@ -130,7 +131,7 @@ router.get("/:userId/inventory/:bookId", async (req, res) => {
     const inventory = await db("books")
       .where({ userId: req.params.userId })
       .first();
-    const book = await Books.getBookById(req.params.bookId).first();
+    const book = await CheckedOut.getBookById(req.params.bookId).first();
     console.log(book);
     if (inventory && book) {
       res.status(200).json(book);
